Stabilizing changes for version 8.4.4 PHP and MySQL 8.0

+++ admin/cp-classes/universal.class.php
@@ -275,7 +275,7 @@ function get_date()
 function get_current_date()
 {
     global $db;
-    $change      = $db->get_option('time_change') * 3600;
+    $change      = (int)$db->get_option('time_change') * 3600;
     $change_time = time() + $change;
 
     return date('Y-m-d H:i:s', $change_time);
@@ -766,12 +766,12 @@ function add_time_to_expires($timeframe, $expires_date = '', $threshold_date = '
             $months_with_31 = array('01', '03', '05', '07', '08', '10', '12');
             $months_with_30 = array('04', '06', '09', '11');
             $leap_years     = array('2008', '2012', '2016', '2020', '2024', '2028');
-            $years          = substr($timeframe, 0, 2);
-            $months         = substr($timeframe, 2, 2);
-            $days           = substr($timeframe, 4, 2);
-            $hours          = substr($timeframe, 6, 2);
-            $minutes        = substr($timeframe, 8, 2);
-            $seconds        = substr($timeframe, 10, 2);
+            $years          = (int)substr($timeframe, 0, 2);
+            $months         = (int)substr($timeframe, 2, 2);
+            $days           = (int)substr($timeframe, 4, 2);
+            $hours          = (int)substr($timeframe, 6, 2);
+            $minutes        = (int)substr($timeframe, 8, 2);
+            $seconds        = (int)substr($timeframe, 10, 2);
             $all_together   = $years . $days . $hours . $minutes . $seconds;
             if ($months > 0 && $all_together == "0") {
                 $expires_split = explode(' ', $expires_date);
